refactor: consolidate the m/s to km/h conversion (CW-406) (#515)
