+++ server/utils/workoutExplorer.ts
@@ -1,57 +1,12 @@
-export const workoutExplorerMetricLabels: Record<string, string> = {
-  durationSec: 'Duration',
-  elapsedTimeSec: 'Elapsed Time',
-  distanceMeters: 'Distance',
-  elevationGain: 'Elevation Gain',
-  trainingLoad: 'Training Load',
-  tss: 'TSS',
-  kilojoules: 'Kilojoules',
-  calories: 'Calories',
-  averageWatts: 'Average Power',
-  maxWatts: 'Max Power',
-  normalizedPower: 'Normalized Power',
-  averageHr: 'Average HR',
-  maxHr: 'Max HR',
-  averageCadence: 'Average Cadence',
-  averageSpeed: 'Average Speed',
-  intensity: 'Intensity',
-  efficiencyFactor: 'Efficiency Factor',
-  decoupling: 'Decoupling',
-  powerHrRatio: 'Power / HR Ratio',
-  variabilityIndex: 'Variability Index',
-  trimp: 'TRIMP',
-  hrLoad: 'HR Load',
-  workAboveFtp: 'Work Above FTP'
-}
+import { metresPerSecondToKmh } from '../../shared/units'
+import {
+  workoutExplorerMetricLabels,
+  workoutExplorerMetricUnits
+} from '../../shared/workout-explorer-metrics'
 
-export const workoutExplorerMetricUnits: Record<string, string> = {
-  durationSec: 'duration',
-  elapsedTimeSec: 'duration',
-  distanceMeters: 'm',
-  elevationGain: 'm',
-  trainingLoad: 'load',
-  tss: 'tss',
-  kilojoules: 'kJ',
-  calories: 'kcal',
-  averageWatts: 'W',
-  maxWatts: 'W',
-  normalizedPower: 'W',
-  averageHr: 'bpm',
-  maxHr: 'bpm',
-  averageCadence: 'rpm',
-  // Correct as-is: `Workout.averageSpeed` is stored in m/s, but
-  // `normalizeWorkoutMetricValue` converts it to km/h before it ever reaches a consumer,
-  // so the unit that goes with the emitted value really is km/h (CW-382).
-  averageSpeed: 'km/h',
-  intensity: '',
-  efficiencyFactor: '',
-  decoupling: '%',
-  powerHrRatio: '',
-  variabilityIndex: '',
-  trimp: 'load',
-  hrLoad: 'load',
-  workAboveFtp: 'kJ'
-}
+// The label/unit maps live in `shared/` so the explorer page can consume the
+// same source (`#shared/workout-explorer-metrics`) instead of keeping a copy.
+export { workoutExplorerMetricLabels, workoutExplorerMetricUnits }
 
 export const allowedWorkoutExplorerSummaryMetrics = new Set(
   Object.keys(workoutExplorerMetricLabels)
@@ -62,6 +17,6 @@ export function normalizeWorkoutMetricValue(field: string, value: unknown) {
   const numeric = Number(value)
   if (Number.isNaN(numeric)) return null
   // averageSpeed is persisted in m/s; the explorer charts it in km/h.
-  if (field === 'averageSpeed') return Number((numeric * 3.6).toFixed(1))
+  if (field === 'averageSpeed') return Number(metresPerSecondToKmh(numeric).toFixed(1))
   return numeric
 }
